requirements.txt
einops
jaxtyping
beartype==0.15.0
numpy
tabulate
scipy
tqdm
omegaconf
wandb
omegaconf

# PyTorch-related library
torchinfo

# Configuration library
hydra-core


# Graphing library
matplotlib
plotly
imageio

# Formatter
ruff
black

# Testing
pytest

# Jupyter notebook
jupyterlab
ipywidgets

# Downloading data
# objaverse

# Processing point cloud
laspy[lazrs,laszip]

# Processing mesh
trimesh
pyglet

# Processing voxel
PyMCubes